+++ packages/playground/blueprints/src/lib/steps/apply-wordpress-patches/index.ts
@@ -21,7 +21,6 @@ export interface ApplyWordPressPatchesStep {
 	patchSecrets?: boolean;
 	disableSiteHealth?: boolean;
 	disableWpNewBlogNotification?: boolean;
-	makeEditorFrameControlled?: boolean;
 	prepareForRunningInsideWebBrowser?: boolean;
 	addFetchNetworkTransport?: boolean;
 }
@@ -49,12 +48,6 @@ export const applyWordPressPatches: StepHandler<
 	}
 	if (options.disableWpNewBlogNotification === true) {
 		await patch.disableWpNewBlogNotification();
-	}
-	if (options.makeEditorFrameControlled === true) {
-		await makeEditorFrameControlled(php, patch.wordpressPath, [
-			`${patch.wordpressPath}/wp-includes/js/dist/block-editor.js`,
-			`${patch.wordpressPath}/wp-includes/js/dist/block-editor.min.js`,
-		]);
 	}
 	if (options.prepareForRunningInsideWebBrowser === true) {
 		await patch.prepareForRunningInsideWebBrowser();
@@ -195,128 +188,4 @@ function randomString(length: number) {
 	for (let i = length; i > 0; --i)
 		result += chars[Math.floor(Math.random() * chars.length)];
 	return result;
-}
-
-export async function applyGutenbergPatchOnce(playground: UniversalPHP) {
-	/**
-	 * Ensures the block editor iframe is controlled by the playground
-	 * service worker. Tl;dr it must use a HTTP URL as its src, not a
-	 * data URL, blob URL, or a srcDoc like it does by default.
-	 *
-	 * @see https://github.com/WordPress/wordpress-playground/pull/668
-	 */
-	const documentRoot = await playground.documentRoot;
-	if (
-		(await playground.isDir(
-			documentRoot + '/wp-content/plugins/gutenberg'
-		)) &&
-		!(await playground.fileExists(documentRoot + '/.gutenberg-patched'))
-	) {
-		await playground.writeFile(documentRoot + '/.gutenberg-patched', '1');
-		await makeEditorFrameControlled(playground, documentRoot, [
-			`${documentRoot}/wp-content/plugins/gutenberg/build/block-editor/index.js`,
-			`${documentRoot}/wp-content/plugins/gutenberg/build/block-editor/index.min.js`,
-		]);
-	}
-}
-
-/**
- *
- * Pair the site editor's nested iframe to the Service Worker.
- *
- * Without the patch below, the site editor initiates network requests that
- * aren't routed through the service worker. That's a known browser issue:
- *
- * * https://bugs.chromium.org/p/chromium/issues/detail?id=880768
- * * https://bugzilla.mozilla.org/show_bug.cgi?id=1293277
- * * https://github.com/w3c/ServiceWorker/issues/765
- *
- * The problem with iframes using srcDoc and src="about:blank" as they
- * fail to inherit the root site's service worker.
- *
- * Gutenberg loads the site editor using <iframe srcDoc="<!doctype html">
- * to force the standards mode and not the quirks mode:
- *
- * https://github.com/WordPress/gutenberg/pull/38855
- *
- * This commit patches the site editor to achieve the same result via
- * <iframe src="/doctype.html"> and a doctype.html file containing just
- * `<!doctype html>`. This allows the iframe to inherit the service worker
- * and correctly load all the css, js, fonts, images, and other assets.
- *
- * Ideally this issue would be fixed directly in Gutenberg and the patch
- * below would be removed.
- *
- * See https://github.com/WordPress/wordpress-playground/issues/42 for more details
- */
-export async function makeEditorFrameControlled(
-	php: UniversalPHP,
-	wordpressPath: string,
-	blockEditorScripts: string[]
-) {
-	const controlledIframe = `
-	/**
-	 * A synchronous function to read a blob URL as text.
-	 *
-	 * @param {string} url
-	 * @returns {string}
-	 */
-	const __playground_readBlobAsText = function (url) {
-		try {
-		  let xhr = new XMLHttpRequest();
-		  xhr.open('GET', url, false);
-		  xhr.overrideMimeType('text/plain;charset=utf-8');
-		  xhr.send();
-		  return xhr.responseText;
-		} catch(e) {
-		  return '';
-		} finally {
-		  URL.revokeObjectURL(url);
-		}
-	}
-
-	window.__playground_ControlledIframe = window.wp.element.forwardRef(function (props, ref) {
-		const source = window.wp.element.useMemo(function () {
-			if (props.srcDoc) {
-				// WordPress <= 6.2 uses a srcDoc that only contains a doctype.
-				return '/wp-includes/empty.html';
-			} else if (props.src && props.src.startsWith('blob:')) {
-				// WordPress 6.3 uses a blob URL with doctype and a list of static assets.
-				// Let's pass the document content to empty.html and render it there.
-				return '/wp-includes/empty.html#' + encodeURIComponent(__playground_readBlobAsText(props.src));
-			} else {
-				// WordPress >= 6.4 uses a plain HTTPS URL that needs no correction.
-				return props.src;
-			}
-		}, [props.src]);
-		return (
-			window.wp.element.createElement('iframe', {
-				...props,
-				ref: ref,
-				src: source,
-				// Make sure there's no srcDoc, as it would interfere with the src.
-				srcDoc: undefined
-			})
-		)
-	});`;
-
-	for (const filePath of blockEditorScripts) {
-		if (!(await php.fileExists(filePath))) {
-			continue;
-		}
-		await updateFile(
-			php,
-			filePath,
-			// The original version of this function crashes WASM PHP, let's define an empty one instead.
-			(contents) =>
-				`${controlledIframe} ${contents.replace(
-					/\(\s*"iframe",/,
-					'(__playground_ControlledIframe,'
-				)}`
-		);
-	}
-	await php.writeFile(
-		`${wordpressPath}/wp-includes/empty.html`,
-		'<!doctype html><script>const hash = window.location.hash.substring(1); if ( hash ) document.write(decodeURIComponent(hash))</script>'
-	);
 }
