--- packages/playground/website/cypress/fixtures/example.json
@@ -1,5 +1,0 @@
-{
-	"name": "Using fixtures to represent data",
-	"email": "hello@cypress.io",
-	"body": "Fixtures are a great way to mock data for responses to routes"
-}
