+++ packages/playground/website/project.json
@@ -112,7 +112,7 @@
 				"cypressConfig": "packages/playground/website/cypress.config.ts",
 				"testingType": "e2e",
 				"devServerTarget": "playground-website:preview",
-				"browser": "chrome"
+				"browser": "chromium"
 			},
 			"dependsOn": ["build"]
 		},
