+++ packages/playground/website/public/wordpress.html
@@ -242,7 +242,6 @@
 						wordpressPath: '/wordpress',
 						addPhpInfo: true,
 						disableSiteHealth: true,
-						makeEditorFrameControlled: true,
 						prepareForRunningInsideWebBrowser: true,
 					},
 					{
